refactor: split the jumpdest scan into std and zkevm files
The two `#if ZK_EVM` regions become a `JumpDestinationAnalyzer.std.cs` /
`.zkevm.cs` pair, which is how the rest of the repo carries a per-flavour
body. The word scanner and its helpers now leave the guest compilation
structurally rather than by relying on ILC to trim them.

Also document the two invariants the review asked for: the positional
contract of `_byteScanThresholds`, and that a tail PUSH can leave the
moving reference past the end of the buffer, where it is only compared.
